[Profiler][Python tracer] Add ephemeral inputs to the value cache.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81958

There are a couple of bugs in the python tracer related to how we cache values. The first is that `ValueCache::store<CallType::PyModuleCall>` wrongly assumes that it will only be called from the profiling callback and calls `PyEval_GetFrame`, effectively violating the encapsulation of the cache by accessing global state. Secondly, we use `arg` to cache bound C functions. This turns out not to be correct, and collisions are resulting in incorrect traces.

In both cases, we can solve the problem by introducing a concept of ephemeral data which is used to materialize a cached value, but is not part of the cache key. (And the author is responsible for making sure that is done correctly.)
ghstack-source-id: 162305160

Differential Revision: [D38062921](https://our.internmc.facebook.com/intern/diff/D38062921/)